moved all files to their respective folders
