+++ src/components/clickable-attribute/ClickableAttribute.js
