+++ src/pages/productView/ProductView.js
@@ -1,8 +1,8 @@
 import React, { PureComponent } from "react";
 import { Query } from "react-apollo";
-import HeaderContainer from "../components/HeaderContainer";
-import ProductContentContainer from "../components/ProductContentContainer";
-import { product_view_query } from "../lib/queries";
+import HeaderContainer from "../../components/header/HeaderContainer";
+import ProductContentContainer from "../../components/product-content/ProductContentContainer";
+import { product_view_query } from "../../lib/queries";
 
 export default class ProductView extends PureComponent {
   constructor(props) {
